refactor: 마일스톤 획득 내역 페이지 내 컴포넌트 분리
#25
