+++ frontend/src/app/(withSidebar)/my-page/milestone/components/MilestoneDetail/styled.ts
