+++ frontend/src/app/(withSidebar)/my-page/milestone/components/MilestoneOverview/styled.ts
@@ -1,0 +1,10 @@
+'use client';
+
+import styled from 'styled-components';
+
+export const MilestoneWrapper = styled.div`
+  min-width: 330px;
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+`;
